add: some small updates

+++ libs/setup-android-libs.sh
@@ -3,22 +3,28 @@ mkdir -p deps; cd ./deps
 mkdir -p downloaded; cd ./downloaded;
 
 set KEYWORD android
-for data in (curl -s https://api.github.com/repos/Qv2ray/Qv2ray-deps/releases/latest | jq ".assets[] | {browser_download_url, name}" -c |  grep $KEYWORD)
-    set NAME (echo $data | jq ".name" -r)
-    echo "Downloading: $NAME"
-    curl -sL (echo $data | jq ".browser_download_url" -r) -o $NAME;
+
+if test -f downloaded_$KEYWORD
+    echo "Download cache for $KEYWORD found, skipping."
+else
+    for data in (curl -s https://api.github.com/repos/Qv2ray/Qv2ray-deps/releases/latest | jq ".assets[] | {browser_download_url, name}" -c |  grep $KEYWORD)
+        set NAME (echo $data | jq ".name" -r)
+        echo "Downloading: $NAME"
+        curl -sL (echo $data | jq ".browser_download_url" -r) -o $NAME;
+    end
+    touch downloaded_$KEYWORD
 end
+
 cd ..
 
-for f in (ls)
-    7z x -y ./$f
+for f in (ls ./downloaded | grep $KEYWORD)
+    7z x -y ./downloaded/$f
 end
 
-mkdir final
 for p in arm arm64 x64 x86
-    cp -rv ./$KEYWORD-$p/installed/* ./final
+    echo "Cleaning up $p-$KEYWORD"
+    rm -rvf ../$p-$KEYWORD/*
+    cp -rvf ./$KEYWORD-$p/installed/$p-$KEYWORD/* ../$p-$KEYWORD
+    rm -rvf ./$KEYWORD-$p/
 end
-
-rm -rvf ./final/vcpkg
 cd ..
-cp -rvf ./deps/final/* ./
