+++ libs/setup-android-tools.sh
@@ -3,10 +3,16 @@ mkdir -p deps; cd ./deps
 mkdir -p downloaded; cd ./downloaded;
 
 set KEYWORD tools
-for data in (curl -s https://api.github.com/repos/Qv2ray/Qv2ray-deps/releases/latest | jq ".assets[] | {browser_download_url, name}" -c |  grep $KEYWORD)
-    set NAME (echo $data | jq ".name" -r)
-    echo "Downloading: $NAME"
-    curl -sL (echo $data | jq ".browser_download_url" -r) -o $NAME;
+
+if test -f downloaded_$KEYWORD
+    echo "Download cache for $KEYWORD found, skipping."
+else
+    for data in (curl -s https://api.github.com/repos/Qv2ray/Qv2ray-deps/releases/latest | jq ".assets[] | {browser_download_url, name}" -c |  grep $KEYWORD)
+        set NAME (echo $data | jq ".name" -r)
+        echo "Downloading: $NAME"
+        curl -sL (echo $data | jq ".browser_download_url" -r) -o $NAME;
+    end
+    touch downloaded_$KEYWORD
 end
 cd ..
 
@@ -15,3 +21,4 @@ for f in (ls ./downloaded | grep $KEYWORD)
 end
 
 cp -rvf ./tools ../
+rm -rvf ./tools
